fix empty struct validation and reuse /core/Predicate module for all refinements
