+++ .changeset/sharp-cooks-glow.md
@@ -1,0 +1,5 @@
+---
+"@fp-ts/schema": patch
+---
+
+fix empty struct validation and reuse /core/Predicate module for all refinements
