+++ src/data/ReadonlyMap.ts
@@ -11,8 +11,7 @@ import * as PR from "@fp-ts/schema/ParseResult"
 import type { Pretty } from "@fp-ts/schema/Pretty"
 import type { Schema } from "@fp-ts/schema/Schema"
 
-const isMap = (u: unknown): u is Map<unknown, unknown> =>
-  typeof u === "object" && typeof u !== null && u instanceof Map
+const isMap = (u: unknown): u is Map<unknown, unknown> => u instanceof Map
 
 const parser = <K, V>(
   key: P.Parser<K>,
